Complete overhaul (#51)
* add a failing test for prerelease increment

* add precommit definition

* add precommit hooks file

* move to semver library

to support incrementing prerelease and build flags

* move semver requirement to requirements.txt

* fix a couple of issues (conflicting versions check was wrong)

add a couple of extra tests
have to clear the config as it's operating as a global with mutable dicts - fine in CLI but breaks in tests

* support for detecting newsfiles added since last release

* add ability to extract the names of file triggers

* allow multiple sources

gives better handling of situation where user is starting from `0.0.0` with no tags.

* fix bug where git was returning files relative to project root

whereas glob returns files relative to cwd

* major: invert news trigger config to allow many-one

example config for `trigger_patterns`:
```
                "newsfiles/*.bugfix": "patch",
                "newsfiles/*.doc": "patch",
                "newsfiles/*.feature": "minor",
                "newsfiles/*.major": "major",
                "newsfiles/*.minor": "minor",
                "newsfiles/*.patch": "patch",
                "newsfiles/*.removal": "major",
```

* major: invert news trigger config to allow many-one

example config for `trigger_patterns`:
```
                "newsfiles/*.bugfix": "patch",
                "newsfiles/*.doc": "patch",
                "newsfiles/*.feature": "minor",
                "newsfiles/*.major": "major",
                "newsfiles/*.minor": "minor",
                "newsfiles/*.patch": "patch",
                "newsfiles/*.removal": "major",
```

* provide a persist-from for previous release

* add functions for comparing semvers

* oops, fix test for trigger config

* add previous-release comparison logic

* oops, fix test for trigger config

* tests and code for correctly incrementing, given a previous release

* use the right param name in the cli

* log which previous release was detected

* minimum patch increment for an existing version

* add support for yaml/yml

* better yaml

* don't perform replacement if line key isn't used

* 'ancestor' tag lookup is more accurate

changes the cli

* match the 'incr_from_release' to the vcs mode (previous or latest)

* move update strings into main. print the right version to stdout.

Previously, the `new_version` was printed, which included prerelease tags, even if the `--release` flag was passed. This seemed wrong, but does change the behaviour - if the prerelease version is needed on stdout, don't pass the `--release` flag.

* Rename inv-conf.major to 3.major

* add git config for ci test run

yuk so much duplication in this file

* fix the tag replacement placeholder

previously, tag templates like `v{version}` would end up looking like

`^(.*)v$`

Because the `10*v` is position-invariant and `.replace` replaces the first match. Using a position-specific string (like, any string with a variety of characters ...) resolves this issue.

* add tests and further improve the regex

* fix the tag replacement placeholder

previously, tag templates like `v{version}` would end up looking like

`^(.*)v$`

Because the `10*v` is position-invariant and `.replace` replaces the first match. Using a position-specific string (like, any string with a variety of characters ...) resolves this issue.

* add tests and further improve the regex

Co-authored-by: David Hyman <david.hyman@five.ai>
Co-authored-by: David Hyman <5198267+davidhyman@users.noreply.github.com>